+++ src/main/scala/sc4pac/ChannelUtil.scala
@@ -117,7 +117,7 @@ object ChannelUtil {
       }
 
       // create symlinks for latest versions
-      channel.contents.foreach { item =>
+      (channel.packages.iterator ++ channel.assets).foreach { item =>
         val latest = item.versions.map(coursier.core.Version(_)).max
         val link = tempJsonDir / MetadataRepository.latestSubPath(item.group, item.name)
         val linkTarget = os.rel / latest.repr
@@ -153,7 +153,7 @@ object ChannelUtil {
             s"""Failed to overwrite folder ${outputDir / "metadata"}. If the problem persists, delete it manually and try again.""")
       }
       os.move.over(tempJsonDir / JsonRepoUtil.channelContentsFilename, outputDir / JsonRepoUtil.channelContentsFilename, createFolders = true)
-      System.err.println(s"Successfully wrote channel contents of ${channel.contents.size} packages and assets.")
+      System.err.println(s"Successfully wrote channel contents of ${channel.packages.size} packages and ${channel.assets.size} assets.")
     })
 
     val packagesTask: RIO[JD.Channel.Info, Seq[JD.PackageAsset]] =
@@ -266,7 +266,10 @@ trait ChannelBuilder[+E] {
         extPkgs.sortBy(i => (i.group, i.name)),
         extAssets.sortBy(i => i.assetId),
       )
-      channel.copy(contents = channel.contents.sortBy(item => (item.group, item.name)))
+      channel.copy(
+        packages = channel.packages.sortBy(item => (item.group, item.name)),
+        assets = channel.assets.sortBy(item => (item.group, item.name)),
+      )
     }
   }
 }